+++ dsmr_backup/services/backup.py
@@ -55,7 +55,7 @@ def create():
     if connection.vendor == 'postgresql':  # pragma: no cover
         backup_process = subprocess.Popen(
             [
-                'pg_dump',
+                settings.DSMRREADER_BACKUP_PG_DUMP,
                 '--host={}'.format(settings.DATABASES['default']['HOST']),
                 '--user={}'.format(settings.DATABASES['default']['USER']),
                 '--dbname={}'.format(settings.DATABASES['default']['NAME']),
@@ -68,7 +68,7 @@ def create():
     elif connection.vendor == 'mysql':  # pragma: no cover
         backup_process = subprocess.Popen(
             [
-                'mysqldump',
+                settings.DSMRREADER_BACKUP_MYSQLDUMP,
                 '--compress',
                 '--hex-blob',
                 '--extended-insert',
@@ -84,7 +84,7 @@ def create():
     elif connection.vendor == 'sqlite':  # pragma: no cover
         backup_process = subprocess.Popen(
             [
-                'sqlite3',
+                settings.DSMRREADER_BACKUP_SQLITE,
                 settings.DATABASES['default']['NAME'],
                 '.dump',
             ],